Move contact-list to its own component
