+++ ui/app/components/app/contact-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './contact-list.component'
