+++ ui/app/components/app/contact-list/recipient-group/index.js
@@ -1,0 +1,1 @@
+export { default } from './recipient-group.component'
